requirements.txt
aiohttp>3.8
async-timeout>=4.0.0a3
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
dateparser==1.0.0
idna==3.2
multidict==5.1.0
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
#regex>2021.7.6
requests==2.26.0
six==1.16.0
tradingview-ta==3.2.4
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
websockets==9.1
yarl<2.0
numpy~=1.21.2
prettytable==2.2.1
bt==0.2.9
pandas-ta==0.3.14b
regex==2022.3.15
ccxt==1.76.38
ta-lib
